Move ReflectionUtils to a separate package
