+++ java-snapshot-testing-core/src/main/java/au/com/origin/snapshots/utils/ReflectionUtils.java
@@ -1,4 +1,4 @@
-package au.com.origin.snapshots;
+package au.com.origin.snapshots.utils;
 
 import java.lang.reflect.Field;
 import java.lang.reflect.Modifier;
